UserService: Implemented createUser and loadUser methods with passing tests. WIP: deleteUser and updateUser methods.

+++ src/main/java/fr/tp/eni/encheresskyjo/bll/impl/UserServiceImpl.java
@@ -3,10 +3,15 @@ package fr.tp.eni.encheresskyjo.bll.impl;
 import fr.tp.eni.encheresskyjo.bll.UserService;
 import fr.tp.eni.encheresskyjo.bo.User;
 import fr.tp.eni.encheresskyjo.converter.UserCreateDtoToUserConverter;
+import fr.tp.eni.encheresskyjo.converter.UserToUserGeneralDTOConverter;
+import fr.tp.eni.encheresskyjo.converter.UserUpdateDtoToUserConverter;
 import fr.tp.eni.encheresskyjo.dal.UserDAO;
 import fr.tp.eni.encheresskyjo.dto.UserCreateDTO;
+import fr.tp.eni.encheresskyjo.dto.UserGeneralDTO;
+import fr.tp.eni.encheresskyjo.dto.UserUpdateDTO;
 import fr.tp.eni.encheresskyjo.exception.BusinessCode;
 import fr.tp.eni.encheresskyjo.exception.BusinessException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -17,12 +22,23 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class UserServiceImpl implements UserService {
-    private UserDAO userDAO;
-    private UserCreateDtoToUserConverter userCreateDtoToUserConverter;
 
-    public UserServiceImpl(UserDAO userDAO, UserCreateDtoToUserConverter userCreateDtoToUserConverter) {
+    private final UserDAO userDAO;
+    private final UserToUserGeneralDTOConverter userToUserGeneralDTOConverter;
+    private final UserCreateDtoToUserConverter userCreateDtoToUserConverter;
+    private final UserUpdateDtoToUserConverter userUpdateDtoToUserConverter;
+
+    public UserServiceImpl(
+            UserDAO userDAO,
+            UserToUserGeneralDTOConverter userToUserGeneralDTOConverter,
+            UserCreateDtoToUserConverter userCreateDtoToUserConverter,
+            UserUpdateDtoToUserConverter userUpdateDtoToUserConverter
+            ) {
         this.userDAO = userDAO;
+        this.userToUserGeneralDTOConverter = userToUserGeneralDTOConverter;
         this.userCreateDtoToUserConverter = userCreateDtoToUserConverter;
+        this.userUpdateDtoToUserConverter = userUpdateDtoToUserConverter;
+
     }
 
     /**
@@ -44,7 +60,7 @@ public class UserServiceImpl implements UserService {
     @Override
     public void createUser(UserCreateDTO dto) {
         BusinessException businessException = new BusinessException();
-        boolean isValid = validateUser(dto, businessException);
+        boolean isValid = validateCreateUser(dto, businessException);
 
         if (!isValid) {
             businessException.addKey(BusinessCode.VALID_USER);
@@ -62,30 +78,60 @@ public class UserServiceImpl implements UserService {
     }
 
     /**
-     *
-     * @param dto
+     * Updates an existing user's information.
+     * @param dto the UserUpdateDTO containing updated user data.
+     * @throws BusinessException if validation fails.
      */
     @Override
     @Transactional
-    public void updateUser(UserCreateDTO dto) {
-        //User existingUser = userDAO.readById(user.getUserId());
-
+    public void updateUser(UserUpdateDTO dto) {
 
         BusinessException businessException = new BusinessException();
-        boolean isValid = validateUser(dto, businessException);
 
-        // TODO : converter UserInputDTO -> User
+        // TODO : with Spring Security, validate current Password against the authenticated user.
+        boolean isValid = validateUpdateUser(dto, businessException);
 
-//        if (isValid) {
-//            this.userDAO.updateAll(dto);
+        if (!isValid) {
+            businessException.addKey(BusinessCode.VALID_USER);
+            throw businessException;
+        }
+
+        User existingUser = userDAO.readById(dto.getId());
+        if (dto.getUsername() != null) existingUser.setUsername(dto.getUsername());
+        if (dto.getLastName() != null) existingUser.setLastName(dto.getLastName());
+        if (dto.getFirstName() != null) existingUser.setFirstName(dto.getFirstName());
+        if (dto.getEmail() != null) existingUser.setEmail(dto.getEmail());
+        if (dto.getTelephone() != null) existingUser.setTelephone(dto.getTelephone());
+        if (dto.getStreet() != null) existingUser.setStreet(dto.getStreet());
+        if (dto.getZip() != null) existingUser.setZip(dto.getZip());
+        if (dto.getCity() != null) existingUser.setCity(dto.getCity());
+
+
+        if (isPasswordChanged(dto)) {
+            // TODO: encode password
+            existingUser.setPassword(dto.getNewPassword());
+        }
+
+        userDAO.updateAll(existingUser);
+
+//        User existingUser = userDAO.readByUsername(dto.getUsername());
+//        User updatedUser = userUpdateDtoToUserConverter.convert(dto);
+//        if (isPasswordChanged(dto)) {
+//            updatedUser.setPassword(dto.getNewPassword());  // Encoder le mot de passe ici
 //        } else {
-//            businessException.addKey(BusinessCode.VALID_USER);
-//            throw businessException;
+//            updatedUser.setPassword(existingUser.getPassword());
 //        }
+//        userDAO.updateAll(updatedUser);
 
     }
 
-    private boolean validateUser(UserCreateDTO dto, BusinessException businessException) {
+    /**
+     * Validates the general user information fields
+     * @param dto the UserGeneralDTO containing the user general information.
+     * @param businessException the object to collect any validation errors.
+     * @return true if all general user fields are valid; false otherwise.
+     */
+    private boolean validateUserGeneral(UserGeneralDTO dto, BusinessException businessException) {
         boolean isValid = true;
 
         isValid = isUsernameValid(dto.getUsername(), businessException);
@@ -96,22 +142,127 @@ public class UserServiceImpl implements UserService {
         isValid &= isStreetValid(dto.getStreet(), businessException);
         isValid &= isZipValid(dto.getZip(), businessException);
         isValid &= isCityValid(dto.getCity(), businessException);
-        isValid &= isPasswordValid(dto.getPassword(), dto.getPasswordConfirm(), businessException);
 
         return isValid;
     }
 
-//  Passer par une DTO :
-    // DAO: readById() -> retourne User
-    // BLL: convertit en DTO
-    // Contrôleur: passe UserDTO dans Formulaire
-
-    // profil d'un autre utilisateur (lien depuis pseudo)
-    @Override
-    public User loadUser(String username) {
-        return userDAO.readByUsername(username);
+    /**
+     * Validate the creation of a user by checking both general user information
+     * and password rules.
+     * @param dto the UserCreateDTO containing the user information to validate.
+     * @param businessException the object to collect any validation errors.
+     * @return true if all fields and password validation pass; false otherwise.
+     */
+    private boolean validateCreateUser(UserCreateDTO dto, BusinessException businessException) {
+        boolean isValid = validateUserGeneral(dto, businessException);
+        isValid &= isPasswordValid(dto.getPassword(), dto.getPasswordConfirm(), businessException);
+        return isValid;
     }
 
+    /**
+     * Validate the user update request, including password change attempt.
+     * @param dto the DTO containing user update information.
+     * @param businessException an exception object used to store validation errors.
+     * @return true if the user update is valid, false otherwise.
+     */
+    private boolean validateUpdateUser(UserUpdateDTO dto, BusinessException businessException) {
+        boolean isValid = true;
+
+        if (dto.getUsername() != null && !dto.getUsername().isBlank()) {
+            isValid &= isUsernameValid(dto.getUsername(), businessException);
+        }
+        if (dto.getLastName() != null && !dto.getLastName().isBlank()) {
+            isValid &= isLastnameValid(dto.getLastName(), businessException);
+        }
+        if (dto.getFirstName() != null && !dto.getFirstName().isBlank()) {
+            isValid &= isFirstnameValid(dto.getFirstName(), businessException);
+        }
+        if (dto.getEmail() != null && !dto.getEmail().isBlank()) {
+            isValid &= isEmailValid(dto.getEmail(), businessException);
+        }
+        if (dto.getTelephone() != null && !dto.getTelephone().isBlank()) {
+            isValid &= isPhoneValid(dto.getTelephone(), businessException);
+        }
+        if (dto.getStreet() != null && !dto.getStreet().isBlank()) {
+            isValid &= isStreetValid(dto.getStreet(), businessException);
+        }
+        if (dto.getZip() != null && !dto.getZip().isBlank()) {
+            isValid &= isZipValid(dto.getZip(), businessException);
+        }
+        if (dto.getCity() != null && !dto.getCity().isBlank()) {
+            isValid &= isCityValid(dto.getCity(), businessException);
+        }
+
+        if (isPasswordChanged(dto)) {
+            // TODO: with Spring Security, validate if the current password matches the one in th DB.
+            // (PasswordEncoder)
+
+            // checks if currentPassword field is provided (TO BE REPLACED BY SPRING SECURITY validation)
+            if (dto.getCurrentPassword() == null || dto.getCurrentPassword().isBlank()) {
+                businessException.addKey(BusinessCode.VALID_USER_PASSWORD_CURRENT_BLANK);
+                isValid = false;
+            }
+
+            // Validate new password and its confirmation.
+            isValid &= isPasswordValid(dto.getNewPassword(), dto.getNewPasswordConfirm(), businessException);
+        }
+
+        return isValid;
+    }
+
+    /**
+     * Checks if the user has attempted to change their password.
+     *
+     * @param dto the DTO containing user update information.
+     * @return true if any of the three password-related fields (current password, new password, confirmation) are not null or not blank.
+     */
+    private boolean isPasswordChanged(UserUpdateDTO dto) {
+
+        if (dto.getCurrentPassword() != null && !dto.getCurrentPassword().isBlank()) {
+            return true;
+        }
+
+        if (dto.getNewPassword() != null && !dto.getNewPassword().isBlank()) {
+            return true;
+        }
+
+        if (dto.getNewPasswordConfirm() != null && !dto.getNewPasswordConfirm().isBlank()) {
+            return true;
+        }
+
+        return false;
+
+    }
+
+
+    /**
+     * Retrieves the profile information of a user by their username.
+     *
+     * <p>
+     *     When a user is logged in, they can view the profiles of other users by clicking on their username.
+     * </p>
+     *
+     * @param username the username of the user whose profile is to be displayed.
+     * @return a UserGeneralDTO containing general profile information of the specified user.
+     */
+    @Override
+    public UserGeneralDTO loadUser(String username) {
+
+        BusinessException businessException = new BusinessException();
+
+        if (!isUsernameValid(username, businessException)) {
+            throw businessException;
+        }
+
+        try {
+            User user = userDAO.readByUsername(username);
+            return userToUserGeneralDTOConverter.convert(user);
+        } catch (EmptyResultDataAccessException e) {
+            businessException.addKey(BusinessCode.USER_NOT_FOUND);
+            throw businessException;
+        }
+
+    }
 
 
     @Override
@@ -120,9 +271,27 @@ public class UserServiceImpl implements UserService {
         userDAO.updatePassword(email, newPassword);
     }
 
-    //TODO
+    /**
+     * Deletes a user account by its ID.
+     *
+     * <p>
+     *     This method is used when a logged-in user chooses to delete their own account,
+     *     or by an administrator to delete another user's account.
+     * </p>
+     *
+     * @param userId the ID of the user to be deleted.
+     */
     @Override
     public void deleteUser(int userId) {
+        // TODO : contraintes de clés étrangères Articles et Enchères
+        //TODO: vérifier userId correspond à user connecté
+
+        User user = userDAO.readById(userId);
+        if (user == null) {
+            BusinessException businessException = new BusinessException();
+            businessException.addKey(BusinessCode.USER_NOT_FOUND);
+            throw businessException;
+        }
         userDAO.delete(userId);
     }
